# api/openapi.yaml
openapi: 3.1.0
info:
  title: go-service-main API
  version: 0.4.0
  description: Single-tenant administration API compatible with vue3-element-admin.
security:
  - bearerAuth: []
paths:
...
  /api/v1/dicts/{dictId}/form:
    parameters:
      - { in: path, name: dictId, required: true, schema: { type: string } }
    get:
      operationId: getDictionaryForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/dicts/{dictId}:
    parameters:
      - { in: path, name: dictId, required: true, schema: { type: string } }
    put:
      operationId: updateDictionary
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteDictionary
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/dicts/{dictCode}/items:
    parameters:
      - { in: path, name: dictCode, required: true, schema: { type: string } }
    get:
      operationId: listDictionaryItems
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    post:
      operationId: createDictionaryItem
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/dicts/{dictCode}/items/options:
    parameters:
      - { in: path, name: dictCode, required: true, schema: { type: string } }
    get:
      operationId: getDictionaryItemOptions
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/dicts/{dictCode}/items/{itemId}/form:
    parameters:
      - { in: path, name: dictCode, required: true, schema: { type: string } }
      - { in: path, name: itemId, required: true, schema: { type: string } }
    get:
      operationId: getDictionaryItemForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/dicts/{dictCode}/items/{itemId}:
    parameters:
      - { in: path, name: dictCode, required: true, schema: { type: string } }
      - { in: path, name: itemId, required: true, schema: { type: string } }
    put:
      operationId: updateDictionaryItem
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteDictionaryItems
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
  /api/v1/files:
    post:
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteFile
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/files/batch:
    post:
      operationId: uploadFiles
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  maxItems: 20
                  items: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/files/image:
    post:
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    AdminAPISuccess:
      description: Successful administration API response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminAPIResult' }
...
  schemas:
...
    AdminAPIResult:
      type: object
      required: [code, msg, data]
      properties:
        code: { type: string }
        msg: { type: string }
        data: {}
        requestId: { type: string }